Most features implemented
