+++ Doberman/Database.py
@@ -1,0 +1,414 @@
+#!/usr/bin/env python3
+import datetime
+import pymongo
+import utils
+import json
+import argparse
+import os
+import time
+import re  # EVERYBODY STAND BACK xkcd.com/208
+dtnow = datetime.datetime.now
+
+__all__ = 'Database'
+
+
+class Database(object):
+    """
+    Class to handle interfacing with the Doberman database
+    """
+
+    def __init__(self, client, appname):
+        self.logger = logging.getLogger(self.__class__.__name__)
+        self.appname = appname
+
+    def close(self):
+        if self.client is not None:
+            self.client.close()
+
+    def __del__(self):
+        self.close()
+        return
+
+    def __exit__(self):
+        self.close()
+        return
+
+    def _check(self, db_name, collection_name):
+        """
+        Returns the requested collection and logs if the database or
+        collection don't yet exist. If the collection must be
+        created, it adds an index onto the 'when' field (used for logs,
+        alarms, data, etc)
+
+        :param db_name: the name of the database
+        :param collection_name: the name of the collections
+        :returns collection instance of the requested collection.
+        """
+        if not hasattr(self, 'experiment_name'):
+            raise ValueError('I don\'t know what experiment to look for')
+        db_name = self.experiment_name + '_' + db_name
+        #if db_name not in self.client.list_database_names():
+        #    self.logger.debug('Database %s doesn\'t exist yet, creating it...' % db_name)
+        if collection_name not in self.client[db_name].list_collection_names(False):
+            self.logger.debug('Collection %s not in database %s, creating it...' % (collection_name, db_name))
+            self.client[db_name].create_collection(collection_name)
+            if 'data' in db_name:
+                self.client[db_name][collection_name].create_index([('when',-1)])
+        return self.client[db_name][collection_name]
+
+    def insertIntoDatabase(self, db_name, collection_name, document, **kwargs):
+        """
+        Inserts document(s) into the specified database/collection
+
+        :param db_name: name of the database
+        :param collection_name: name of the collection
+        :param document: a dictionary or iterable of dictionaries
+        :param **kwargs: any keyword args, passed to collection.insert_(one|many)
+        :returns 0 if successful, -1 if a multiple insert failed, -2 if a single
+        insert failed, or 1 if `document` has the wrong type
+        """
+        collection = self._check(db_name, collection_name)
+        if isinstance(document, (list, tuple)):
+            result = collection.insert_many(document, **kwargs)
+            if len(result.inserted_ids) != len(document):
+                self.logger.error('Inserted %i entries instead of %i into %s/%s' % (
+                    len(result.inserted_ids), len(document), db_name, collection_name))
+                return -1
+            return 0
+        elif isinstance(document, dict):
+            result = collection.insert_one(document, **kwargs)
+            if result.acknowledged:
+                return 0
+            else:
+                return -2
+        else:
+            self.logger.error('Not sure what to do with %s type' % type(document))
+            return 1
+
+    def readFromDatabase(self, db_name, collection_name, cuts={}, onlyone = False, **kwargs):
+        """
+        Finds one or more documents that pass the specified cuts
+
+        :param db_name: name of the database
+        :param collection_name: name of the collection
+        :param cuts: dictionary of the query to apply. Default {}
+        :param onlyone: bool, if only one document is requested
+        :param **kwargs: keyword args passed to collection.find. The 'sort' kwarg
+        is handled separately because otherwise it doesn't do anything
+        :returns document if onlyone=True else cursor
+        """
+        collection = self._check(db_name,collection_name)
+        cursor = collection.find(cuts, **kwargs)
+        if 'sort' in kwargs:
+            cursor.sort(kwargs['sort'])
+        if onlyone:
+            for doc in cursor:
+                return doc
+        else:
+            return cursor
+
+    def updateDatabase(self, db_name, collection_name, cuts, updates, **kwargs):
+        """
+        Updates documents that meet pass the specified cuts
+
+        :param db_name: name of the database
+        :param collection_name: name of the collection
+        :param cuts: the dictionary specifying the query
+        :param updates: the dictionary specifying the desired changes
+        :param **kwargs: keyword args passed to collection.update_many
+        :returns 0 if successful, 1 if not
+        """
+        collection = self._check(db_name, collection_name)
+        ret = collection.update_many(cuts, updates, **kwargs)
+        if not ret.acknowledged:
+            return 1
+        return 0
+
+    def DeleteDocuments(self, db_name, collection_name, cuts):
+        """
+        Deletes documents from the specified collection
+
+        :param db_name: name of the database
+        :param collection_name: name of the collection
+        :param cuts: dictionary specifying the query
+        :returns None
+        """
+        collection = self._check(db_name, collection_name)
+        collection.delete_many(cuts)
+
+    def Distinct(self, db_name, collection_name, field, cuts={}, **kwargs):
+        """
+        Transfer function for collection.distinct
+        """
+        return self._check(db_name, collection_name).distinct(field, cuts, **kwargs)
+
+    def Count(self, db_name, collection_name, cuts, **kwargs):
+        """
+        Transfer function for collection.count/count_documents
+        """
+        return self._check(db_name, collection_name).count_documents(cuts, **kwargs)
+
+    def FindOneAndUpdate(self, db_name, collection_name, cuts, updates, **kwargs):
+        """
+        Finds one document and applies updates. A bit of a special implementation so
+        the 'sort' kwarg will actually do something
+
+        :param db_name: name of the database
+        :param collection_name: name of the collection
+        :param cuts: a dictionary specifying the query
+        :param updates: a dictionary specifying the updates
+        :**kwargs: keyword args passed to readFromDatabase
+        :returns document
+        """
+        doc = self.readFromDatabase(db_name, collection_name, cuts, onlyone=True, **kwargs)
+        if doc is not None:
+            self.updateDatabase(db_name, collection_name, {'_id' : doc['_id']}, updates)
+        return doc
+
+    def FindCommand(self, name):
+        """
+        Finds the oldest unacknowledged command for the specified entity
+        and updates it as acknowledged. Deletes command documents used in
+        the feedback subsystem
+
+        :param name: the entity to find a command for
+        :returns command document
+        """
+        now = dtnow()
+        doc = self.FindOneAndUpdate('logging', 'commands',
+                cuts={'name' : name,
+                      'acknowledged' : {'$exists' : 0},
+                      'logged' : {'$lte' : now}},
+                updates={'$set' : {'acknowledged' : now}},
+                sort=[('logged',1)])
+        if doc and 'by' in doc and doc['by'] == 'feedback':
+            self.DeleteDocuments('logging', 'commands', {'_id' : doc['_id']})
+        return doc
+
+    def getMessageProtocols(self, level):
+        """
+        Gets message protocols for the specified alarm level. If none are found,
+        takes those from the next lowest level.
+
+        :param level: which alarm level is in question (0, 1, etc)
+        :returns: list of message protocols to use
+        """
+        doc = self.readFromDatabase('settings','alarm_config', {'level' : level}, onlyone=True)
+        if doc is None:
+            self.logger.error('No message protocols for alarm level %i! Defaulting to next lowest level' % level)
+            doc = self.readFromDatabase('settings','alarm_config',
+                    {'level' : {'$lte' : level}}, onlyone=True,
+                    sort=[('level', -1)])
+        return doc['protocols']
+
+    def getContactAddresses(self, level):
+        """
+        Returns a list of addresses to contact at 'level'
+
+        :param level: which alarm level the message will be sent at
+        :returns dict, keys = message protocols, values = list of addresses
+        """
+        protocols = self.getMessageProtocols(level)
+        ret = {k : [] for k in protocols}
+        for doc in self.readFromDatabase('settings','contacts',
+                {'status' : {'$gte' : 0, '$lte' : level}}):
+            for p in protocols:
+                ret[p].append(doc[p])
+        return ret
+
+    def UpdateHeartbeat(self, host=None, sensor=None):
+        """
+        Heartbeats the specified sensor or host
+        """
+        if host is not None:
+            cuts = {'host' : host}
+            coll = 'hosts'
+        elif sensor is not None:
+            cuts = {'sensor' : sensor}
+            coll = 'sensors'
+        self.updateDatabase('settings', coll, cuts=cuts,
+                    updates={'$set' : {'heartbeat' : dtnow()}})
+        return
+
+    def CheckHeartbeat(self, name):
+        """
+        Checks the heartbeat of the specified sensor.
+
+        :param name: the name of the sensor to check
+        :returns: number of seconds since the last heartbeat
+        """
+        doc = self.GetSensorSettings(name=name)
+        last_heartbeat = doc['heartbeat']
+        return (dtnow() - last_heartbeat).total_seconds()
+
+    def logAlarm(self, document):
+        """
+        Adds the alarm to the history.
+        """
+        if self.insertIntoDatabase('logging','alarm_history',document):
+            self.logger.warning('Could not add entry to alarm history!')
+            return -1
+        return 0
+
+    def LogUpdate(self, **kwargs):
+        """
+        Logs changes submitted from the website
+        """
+        self.insertIntoDatabase('logging', 'updates', kwargs)
+        return
+
+    def GetSensorSetting(self, sensor_name, field=None):
+        """
+        Gets a specific setting from one sensor
+
+        :param sensor_name: the name of the sensor
+        :param field: the field you want
+        :returns: the value of the named field
+        """
+        doc = self.readFromDatabase('settings', 'sensors', cuts={'name' : name},
+                onlyone=True)
+        if field is not None:
+            return doc
+        return doc[field]
+
+    def GetReading(self, sensor=None, name=None):
+        """
+        Gets the document from one reading
+
+        :param sensor: the name of the sensor
+        :param name: the name of the reading
+        :returns: reading document
+        """
+        return self.readFromDatabase('settings', 'readings',
+                cuts={'sensor' : sensor, 'name' : name}, onlyone=True)
+
+    def UpdateReading(self, sensor=None, name=None, field=None, value=None):
+        """
+        Updates a parameter for a reading
+
+        :param sensor: the name of the sensor
+        :param name: the name of the reading
+        :param field: the specific field to update
+        :param value: the new value
+        """
+        self.updateDatabase('settings', 'readings',
+                cuts={'sensor' : sensor, 'name' : name},
+                updates={'$set' : {field : value}})
+
+    def SetSensorSetting(self, name, field, value):
+        """
+        Updates the setting from one sensor
+
+        :param name: the name of the sensor
+        :param field: the specific field to update
+        :param value: the new value
+        """
+        self.updateDatabase('settings', 'sensors', cuts={'name' : name},
+                updates = {'$set' : {field : value}})
+
+    def getDefaultSettings(self, runmode=None, name=None):
+        """
+        Reads default Doberman settings from database.
+
+        :param runmode: the runmode to get settings for
+        :param name: the name of the setting
+        :returns: the setting dictionary if name=None, otherwise the specific field
+        """
+        if runmode:
+            doc = self.readFromDatabase('settings','runmodes',
+                    {'mode' : runmode}, onlyone=True)
+            if name:
+                return doc[name]
+            return doc
+        doc = self.readFromDatabase('settings','current_status', onlyone=True)
+        if name:
+            return doc[name]
+        return doc
+
+    def GetHostSetting(self, host=self.hostname, field=None):
+        """
+        Gets the setting document of the specified host
+        """
+        doc = self.readFromDatabase('settings', 'hosts', {'hostname' : host},
+                onlyOne=True)
+        if field is not None and field in doc:
+            return doc[field]
+        return doc
+
+    def SetHostSetting(self, host=self.hostname, **kwargs):
+        """
+        Updates the setting document of the specified host. Kwargs should be one
+        of the Mongo commands (set, unset, push, pull) without the $ char.
+        Ex: set={field:value}
+        """
+        self.updateDatabase('settings', 'hosts', {'host' : host},
+                updates=dict(f'${k}' : v for k,v in kwargs.items()))
+
+    def ManagePlugins(self, name, action):
+        """
+        Adds or removes a plugin from the managed list. Doberman adds, plugins remove
+
+        :param name: the name of the plugin
+        :param action: 'add' or 'remove'
+        :returns: None
+        """
+        managed_plugins = self.getDefaultSettings(name='managed_plugins')
+        if action=='add':
+            if name in managed_plugins:
+                self.logger.info('%s already managed' % name)
+            else:
+                self.updateDatabase('settings','current_status',cuts={},
+                        updates={'$push' : {'managed_plugins' : name}})
+        elif action=='remove':
+            if name not in managed_plugins:
+                self.logger.debug('%s isn\'t managed' % name)
+            else:
+                self.updateDatabase('settings','current_status',cuts={},
+                        updates={'$pull' : {'managed_plugins' : name}})
+        return
+
+    def WriteDataToDatabase(self, sensor_name, data_doc):
+        """
+        Writes data to the database
+
+        :param sensor_name: the name of the sensor
+        :param data_doc: the document with data
+        :returns 0 on success, -1 otherwise
+        """
+        ret = self.insertIntoDatabase('data', sensor_name, data_doc)
+        if ret:
+            self.logger.warning("Can not write data from %s to Database. "
+                                "Database interaction error." % sensor_name)
+            return -1
+        return 0
+
+    def GetCurrentStatus(self):
+        """
+        Gives a snapshot of the current system status
+        """
+        status = {}
+        now = dtnow()
+        for sensor_doc in self.readFromDatabase('settings','sensors'):
+            sensor_name = sensor_doc['name']
+            #if 'Test' in sensor_name:
+            #    continue
+            status[sensor_name] = {
+                    'status' : sensor_doc['status'],
+                    'last_heartbeat' : (now - sensor_doc['heartbeat']).total_seconds(),
+                    'readings' : {}
+                }
+            for reading_name in sensor_doc['readings']:
+                reading_doc = self.GetReading(sensor_name, reading_name)
+                status[sensor_name]['readings'][reading_name] = {
+                        'description' : reading_doc['description'],
+                        'status' : reading_doc['status'],
+                    }
+                if reading_doc['status'] == 'online':
+                    status[sensor_name]['readings'][reading_name]['runmode'] = reading_doc['runmode'],
+                    data_doc = self.readFromDatabase('data', sensor_name,
+                                cuts={reading_name : {'$exists' : 1}},
+                                sort=[('_id', -1)], onlyone=True)
+                    status[sensor_name]['readings'][reading_name]['last_value'] = data_doc[reading_name]
+                    doc_time = int(str(data_doc['_id'])[:8], 16)
+                    status[sensor_name]['readings'][reading_name]['last_time'] = time.time() - doc_time
+        return status
