+++ Doberman/ExampleSensor.py
