+++ Doberman/RevolutionPi.py
@@ -1,0 +1,9 @@
+import Doberman
+
+
+
+class RevolutionPi(Doberman.Sensor):
+    """
+    A class for the rev-pi PLCs
+    """
+    pass
