+++ Doberman/utils.py
@@ -1,4 +1,3 @@
-from ast import literal_eval
 import serial
 from subprocess import Popen, PIPE, TimeoutExpired
 import importlib
@@ -9,121 +8,16 @@ import signal
 import os.path
 import inspect
 import re
+import logging
+import logging.handlers
 dtnow = datetime.datetime.now
 
+__all__ = 'FindPlugin SensorOpts Logger'.split()
 
-heartbeat_timer = 300
+heartbeat_timer = 30
 buffer_timer = 5
 number_regex = r'[\-+]?[0-9]+(?:\.[0-9]+)?(?:[eE][\-+]?[0-9]+)?'
 doberman_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
-
-def getUserInput(text, input_type=None, be_in=None, be_not_in=None, be_array=False, limits=None, string_length=None, exceptions=None):
-    """
-    Ask for an input bye displaying the 'text'.
-    It is asked until:
-      the input has the 'input_type(s)' specified,
-      the input is in the list 'be_in' (if not None),
-      not in the list 'be_not_in' (if not None),
-      the input is between the limits (if not None).
-      has the right length if it is a string (if not None)
-    'input_type', 'be_in' and 'be_not_in' must be lists or None.
-    'limits' must be a list of type [lower_limit, higher_limit].
-    ' lower_limit' or 'higher_limit' can be None. The limit is <=/>=.
-    'string_length' must be a list of [lower_length, higher_length]
-    ' lower_length' or 'higher_length' can be None. The limit is <=/>=.
-    'be_array' can be True or False, it returns the input as array or not.
-    If the input is in the exceptions it is returned without checks.
-    """
-    while True:
-        # Read input.
-        try:
-            user_input = input_eval(input(text), input_type != [str])
-        except KeyboardInterrupt:
-            raise
-        except Exception as e:
-            print("Error: %s. Try again." % e)
-            continue
-        # Check for input exceptions
-        if exceptions:
-            if user_input in exceptions:
-                return user_input
-        # Remove string signs
-        if input_type == [str] and isinstance(user_input, str):
-            user_input = ''.join(c for c in user_input if c not in ['"',"'"])
-        # Transform input to list
-        if not be_array:
-            user_input = [user_input]
-        else:
-            if isinstance(user_input, tuple):
-                user_input = list(user_input)
-            elif isinstance(user_input, str):
-                user_input = user_input.split(",")
-            elif isinstance(user_input, (int, float)):
-                user_input = [user_input]
-        # Remove spaces after comma for input lists
-        if be_array and input_type == [str]:
-            user_input = [item.strip() for item in user_input]
-        # Check input for type, be_in, be_not_in, limits.
-        if input_type:
-            if not all(isinstance(item, tuple(input_type)) for item in user_input):
-                print("Wrong input format. Must be in %s. Try again." %
-                    str(tuple(input_type)))
-                continue
-        if be_in:
-            if any(item not in be_in for item in user_input):
-                print("Input must be in: %s. Try again." % str(be_in))
-                continue
-        if be_not_in:
-            if any(item in be_not_in for item in user_input):
-                print("Input is not allowed to be in: %s. Try again." % str(be_not_in))
-                continue
-        if limits:
-            if limits[0] or limits[0] == 0:  # Allows also 0.0 as lower limit
-                if any(item < limits[0] for item in user_input):
-                    print("Input must be between: %s. Try again." % str(limits))
-                    continue
-            if limits[1]:
-                if any(item > limits[1] for item in user_input):
-                    print("Input must be between: %s. Try again." % str(limits))
-                    continue
-        # Check for string length
-        if string_length:
-            if string_length[0] != None:
-                if any(len(item) < string_length[0] for item in user_input):
-                    print("Input string must have more than %s characters."
-                        " Try again." % str(string_length[0]))
-                    continue
-            if string_length[1] != None:
-                if any(len(item) > string_length[1] for item in user_input):
-                    print("Input string must have less than %s characters."
-                        " Try again." % str(string_length[1]))
-                    continue
-        break
-    if not be_array:
-        return user_input[0]
-    return user_input
-
-def adjustListLength(input_list, length, append_item, input_name=None):
-    """
-    Appending 'append_item' to the 'input_list'
-    until 'length' is reached.
-    """
-    if len(input_list) < length:
-        input_list += [append_item]*(length - len(input_list))
-    elif len(input_list) > length:
-        input_list = input_list[:length]
-    return input_list
-
-def input_eval(inputstr, literaleval=True):
-    if not isinstance(inputstr, str):
-        return -1
-    inputstr = inputstr.strip(' \r\t\n\'"').expandtabs(4)
-    if literaleval:
-        try:
-            return literal_eval(inputstr)
-        except:
-            pass
-    return str(inputstr)
 
 def refreshTTY(db):
     """
@@ -241,23 +135,11 @@ def FindPlugin(name, path):
         raise AttributeError('Cound not find constructor for %s!' % name)
     return sensor_ctor
 
-def SensorOpts(config_doc):
-    """
-    Returns the dictionary of settings to be passed to a Sensor constructor
-    """
-    opts = {}
-    opts['name'] = config_doc['name']
-    opts.update(config_doc['address'])
-    opts['readings'] = config_doc['readings']
-    if 'additional_params' in config_doc:
-        opts.update(config_doc['additional_params'])
-    return opts
-
 class SignalHandler(object):
     """ Handles signals from the OS
     """
     def __init__(self, logger=None):
-        self.interrupted = False
+        self.run = True
         signal.signal(signal.SIGINT, self.interrupt)
         signal.signal(signal.SIGTERM, self.interrupt)
         self.logger = logger
@@ -266,4 +148,59 @@ class SignalHandler(object):
         if self.logger is not None:
             self.logger.info('Received signal %i' % args[0])
         self.signal_number = int(args[0])
-        self.interrupted = True
+        self.run = False
+
+class DobermanLogger(logging.Handler):
+    """
+    Custom logging interface for Doberman. Logs to
+    the database (with disk as backup).
+    """
+    def __init__(self, db, level=logging.INFO):
+        logging.Handler.__init__(self)
+        self.db = db
+        self.db_name = 'logging'
+        self.collection_name = 'logs'
+        backup_filename = datetime.date.today().isoformat()
+        self.backup_logger = logging.handlers.TimedRotatingFileHandler(
+                os.path.join(doberman_dir, 'logs', backup_filename + '.log'),
+                when='midnight', delay=True)
+        self.stream = logging.StreamHandler()
+        f = logging.Formatter('%(asctime)s | '
+                '%(levelname)s | %(name)s | %(funcName)s | '
+                '%(lineno)d | %(message)s')
+        self.setFormatter(f)
+        self.stream.setFormatter(f)
+        self.backup_logger.setFormatter(f)
+        self.level = level
+
+    def close(self):
+        self.backup_logger.close()
+        self.stream.close()
+        self.db = None
+        return
+
+    def __del__(self):
+        self.close()
+        return
+
+    def emit(self, record):
+        self.stream.emit(record)
+        if record.levelno < self.level:
+            return
+        rec = dict(
+                msg         = record.msg,
+                level       = record.levelno,
+                name        = record.name,
+                funcname    = record.funcName,
+                lineno      = record.lineno)
+        if self.db.insertIntoDatabase(self.db_name, self.collection_name, rec):
+            self.backup_logger.emit(record)
+
+def Logger(name, db, loglevel='INFO'):
+    logger = logging.getLogger(name)
+    try:
+        logger.setLevel(getattr(logging, loglevel))
+    except:
+        logger.setLevel(logging.INFO)
+    logger.addHandler(DobermanLogger(db))
+    return logger
